==> 04_get_weather_data.ipynb <==
## Get Weather Data, Web Scraping

# Web scraping, get weather data for a specific geographical location over a 
# given period of time
# https://open-meteo.com/
# https://archive-api.open-meteo.com/v1/era5


# Import Pandas data manipulation and analysis library
import pandas as pd

# Import requests library
# The requests library is one of the most popular Python libraries for making
# HTTP requests (like GET, POST, PUT, DELETE). It allows Python programs to
# send and receive data from web servers
import requests


# A query to Open-Meteo's historical weather API.
# Base URL: https://archive-api.open-meteo.com/v1/era5
# This endpoint provides ERA5 reanalysis weather data.
# Query Parameters:
# latitude=52.52 and longitude=13.41 - these coordinates point to Berlin, Germany.
# start_date=2021-01-01 and end_date=2021-12-31 - request data for the entire
#   year of 2021.
# hourly=temperature_2m - request the hourly temperature at 2 meters above ground.
url = "https://archive-api.open-meteo.com/v1/era5?latitude=52.52&longitude=13.41&start_date=2021-01-01&end_date=2021-12-31&hourly=temperature_2m"

# Send the request and parse JSON response
response = requests.get(url)
response.json()


# A query to Open-Meteo's historical weather API.
# url - base URL to the ERA5 historical weather API endpoint.
# date - date for Weather Data
# params - parameters for the API Request
#   latitude=41.85, longitude=-87.65 - location of Chicago, IL
#   start_date & end_date - limits the data to June 1, 2023
#   hourly - you’re asking for four hourly weather variables
#   temperature_2m - temperature 2 meters above ground
#   windspeed_10m - wind speed at 10 meters
#   rain - hourly rain in millimeters
#   precipitation - total precipitation (rain + snow, etc.)
url = "https://archive-api.open-meteo.com/v1/era5"
date = "2024-06-01"
params = {
    "latitude": 41.85,
    "longitude": -87.65,
    "start_date": date,
    "end_date": date,
    "hourly": "temperature_2m,windspeed_10m,rain,precipitation",
}
response = requests.get(url, params=params)
response.json()


# Extract part
from datetime import datetime
from dateutil.relativedelta import relativedelta
current_datetime = datetime.now() - relativedelta(months=2)
formatted_datetime = current_datetime.strftime("%Y-%m-%d")
url = "https://archive-api.open-meteo.com/v1/era5"
params = {
    "latitude": 41.85,
    "longitude": -87.65,
    "start_date": formatted_datetime,
    "end_date": formatted_datetime,
    "hourly": "temperature_2m,wind_speed_10m,rain,precipitation",
}
response = requests.get(url, params=params)
weather_data = response.json()
weather_data


# Get hourly timestamps from a weather data
weather_data["hourly"]


# Get hourly timestamps and time from a weather data
weather_data["hourly"]["time"]


# Get hourly timestamps and temperature_2m from a weather data
weather_data["hourly"]["temperature_2m"]


# Creating a new dictionary called weather_data_filtered that selects and 
# renames specific fields from the original weather_data
weather_data_filtered = {
    "datetime": weather_data["hourly"]["time"],
    "tempretaure": weather_data["hourly"]["temperature_2m"],
    "wind_speed": weather_data["hourly"]["wind_speed_10m"],
    "rain": weather_data["hourly"]["rain"],
    "precipitation": weather_data["hourly"]["precipitation"],
}
weather_data_filtered


# Convert filtered weather data into a pandas DataFrame and show its 
# first 5 rows
weather_df = pd.DataFrame(weather_data_filtered)
weather_df.head()


# Get a concise summary of the weather_df
weather_df.info()


# Convert the datetime column to proper datetime64[ns] format
weather_df["datetime"] = pd.to_datetime(weather_df["datetime"])
weather_df.head()


# Weather data transformation part alltogether
# - build a filtered weather_data version of original JSON file
# - convert it to a DataFrame
# - parse the datetime columnt to proper datetime64[ns] format
# - export weather_data to CSV

weather_data_filtered = {
    "datetime": weather_data["hourly"]["time"],
    "tempretaure": weather_data["hourly"]["temperature_2m"],
    "wind_speed": weather_data["hourly"]["wind_speed_10m"],
    "rain": weather_data["hourly"]["rain"],
    "precipitation": weather_data["hourly"]["precipitation"],
}

weather_df = pd.DataFrame(weather_data_filtered)

weather_df["datetime"] = pd.to_datetime(weather_df["datetime"])

weather_df.to_csv("weather_data_date.csv", index=False)
