==> 08_local_visualisations.ipynb <==
## Compile Taxi Trips Aggregated Table

# Compile taxi trips aggregated table ftom
# - taxi trips data
# - community areas data 
# - weather data 

# Import StringIO class in the io module that allows to treat a string as if
# it were a file object
from io import StringIO
import os

# Import load_dotenv function from the python-dotenv library
from dotenv import load_dotenv

# Import AWS SDK for Python to interact with AWS services from python code
import boto3

# Import Pandas data manipulation and analysis library
import pandas as pd


# Looks for a .env file (by default, in the current directory), reads it, and
# adds the variables to the environment.
load_dotenv()


# Retrieving sensitive credentials (an AWS Access Key and Secret Key) from
# environment variables using the os module.
aws_access_key_id = os.getenv("AWS_ACCESS_KEY")
aws_secret_key = os.getenv("AWS_SECRET_KEY")


def read_csv_from_s3(bucket: str, path: str, filename: str) -> pd.DataFrame:
    """
    Download cscv file from S3 bucket.

    Parameters:
    bucket (str): The bucket where the files at.
    path (str): The folder to the file.
    filename (str): Original file name.

    Returns:
        pd:DataFrame - the dataframe of the downloaded file
    """
    s3 = boto3.client(
        "s3", aws_access_key_id=aws_access_key_id, aws_secret_access_key=aws_secret_key
    )
    full_path = f"{path}{filename}"
    object = s3.get_object(Bucket=bucket, Key=full_path)
    object = object["Body"].read().decode("utf-8")
    output_df = pd.read_csv(StringIO(object))
    return output_df


# Use boto3 to connect to AWS S3, then reads several CSV files from specific
# folders (or "paths") in an S3 bucket, loading each one into memory using a
# custom function called read_csv_from_s3
s3 = boto3.client(
    "s3", aws_access_key_id=aws_access_key_id, aws_secret_access_key=aws_secret_key
)
bucket = "cubix-chicago-taxi-ke"

community_areas_path = "transformed_data/community_areas/"
company_path = "transformed_data/company/"
date_path = "transformed_data/date/"
payment_type_path = "transformed_data/payment_type/"

# Note: Taxi and wetaher data files processed separately!
taxi_trips_path = "transformed_data/taxi_trips/"
weather_path = "transformed_data/weather/"

community_areas = read_csv_from_s3(
    bucket=bucket, path=community_areas_path, filename="community_areas_master.csv"
)
company = read_csv_from_s3(
    bucket=bucket, path=company_path, filename="company_master.csv"
)
date = read_csv_from_s3(bucket=bucket, path=date_path, filename="weather_data_date.csv")
payment_type = read_csv_from_s3(
    bucket=bucket, path=payment_type_path, filename="payment_type_master.csv"
)


# Process taxi data files
trips_list = []
for file in s3.list_objects(Bucket=bucket, Prefix=taxi_trips_path).get("Contents", []):
    taxi_trip_key = file["Key"]
    filename = taxi_trip_key.split("/")[-1]

    if filename and filename.endswith(".csv"):
        filename = taxi_trip_key.split("/")[-1]

        # print(filename)
        trip = read_csv_from_s3(bucket, taxi_trips_path, filename)
        trips_list.append(trip)
        print(f"{filename} has been added.")
trips = pd.concat(trips_list, ignore_index=True)
trips.head()


# Process weather data files
weather_list = []
for file in s3.list_objects(Bucket=bucket, Prefix=weather_path).get("Contents", []):
    weather_key = file["Key"]
    filename = weather_key.split("/")[-1]

    if filename and filename.endswith(".csv"):
        filename = weather_key.split("/")[-1]
        # print(filename)
        weather_daily = read_csv_from_s3(bucket, weather_path, filename)
        weather_list.append(weather_daily)
        print(f"{filename} has been added.")
weather = pd.concat(weather_list, ignore_index=True)
weather.head()


# Join two DataFrames (trips and weather) based on matching values
trips_full = pd.merge(
    trips, weather, left_on="datetime_for_weather", right_on="datetime", how="inner"
)
trips_full = trips_full.drop(columns=["datetime"])


# Join two DataFrames (trips_full and company) based on matching values
trips_full = pd.merge(
    trips_full, company, left_on="company_id", right_on="company_id", how="inner"
)
trips_full = trips_full.drop(columns=["company_id"])


# Join two DataFrames (trips_full and payment_type) based on matching values
trips_full = pd.merge(
    trips_full,
    payment_type,
    left_on="payment_type_id",
    right_on="payment_type_id",
    how="inner",
)
trips_full = trips_full.drop(columns=["payment_type_id"])


# Join two DataFrames (trips_full and community_areas) based on
# "pickup_community_area_id" matching values
trips_full = pd.merge(
    trips_full,
    community_areas,
    left_on="pickup_community_area_id",
    right_on="area_code",
    how="inner",
)
trips_full = trips_full.drop(columns=["pickup_community_area_id"])
trips_full.rename(
    columns={"community_name": "pickup_community_area_name"}, inplace=True
)


# Join two DataFrames (trips_full and community_areas) based on
# "dropoff_community_area_id" matching values
trips_full = pd.merge(
    trips_full,
    community_areas,
    left_on="dropoff_community_area_id",
    right_on="area_code",
    how="inner",
)
trips_full = trips_full.drop(columns=["dropoff_community_area_id"])
trips_full.rename(
    columns={"community_name": "dropoff_community_area_name"}, inplace=True
)


# Controlling DataFrame display so you can actually see your full data without
# stuff getting cut off
pd.set_option("display.max_columns", 50)
trips_full.head()


# Display date
date


# Convert the datetime column in the date DataFrame to a proper datetime object
date["date"] = pd.to_datetime(date["datetime"])
# Convert the trip_start_timestamp column into a datetime.
trips_full["trip_start_timestamp"] = pd.to_datetime(trips_full["trip_start_timestamp"])
# Extract just the date part from the full timestamp, dropping the time
trips_full["trip_start_date"] = trips_full["trip_start_timestamp"].dt.date
trips_full["trip_start_date"]


# Join two DataFrames (trips_full and date) based on matching values
trips_full = pd.merge(
    trips_full, date, left_on="trip_start_timestamp", right_on="date", how="inner"
)
trips_full = trips_full.drop(columns=["date"])
trips_full.head()


# Save trips_full to a CSV file
trips_full.to_csv("trips_full.csv", index=False)
